Moved DataStore files into data package.
