+++ QuestionApp/src/ca/ualberta/cs/cmput301f14t14/questionapp/data/LocalDataStore.java
@@ -1,4 +1,4 @@
-package ca.ualberta.cs.cmput301f14t14.questionapp;
+package ca.ualberta.cs.cmput301f14t14.questionapp.data;
 
 import java.util.List;
 import java.util.UUID;
